faker/requirements.txt
kafka-python>=2.0.2
python-schema-registry-client>=0.0.6